Fix printf format warnings.
We can't print function pointer addresses at all, since converting
a function pointer to a void* is forbidden with ANSI C, so we only
print the function name.

+++ jsstring.c
@@ -159,7 +159,7 @@ static void Sp_lastIndexOf(js_State *J)
 {
 	const char *haystack = checkstring(J, 0);
 	const char *needle = js_tostring(J, 1);
-	int pos = js_isdefined(J, 2) ? js_tointeger(J, 2) : strlen(haystack);
+	int pos = js_isdefined(J, 2) ? js_tointeger(J, 2) : (int)strlen(haystack);
 	int len = strlen(needle);
 	int k = 0, last = -1;
 	Rune rune;
